refactor: major span tag build reuse
