+++ otel_tags.go
@@ -1,0 +1,24 @@
+package messaging
+
+const tag_messaging_system_key string = "messaging.system"
+const tag_messaging_operation_key string = "messaging.operation"
+const tag_messaging_destination_key string = "messaging.destination"
+const tag_messaging_destination_kind_key string = "messaging.destination_kind"
+const tag_messaging_protocol_key string = "messaging.protocol"
+const tag_messaging_protocol_version_key string = "messaging.protocol_version"
+const tag_messaging_message_id_key string = "messaging.message_id"
+const tag_messaging_message_payload_size_bytes_key string = "messaging.message_payload_size_bytes"
+const tag_messaging_url_key string = "messaging.url"
+const tag_messaging_rabbitmq_routing_key_key string = "messaging.rabbitmq.routing_key"
+
+const tag_net_sock_family_key string = "net.sock.family"
+const tag_net_sock_peer_port_key string = "net.sock.peer.port"
+const tag_net_sock_peer_name_key string = "net.sock.peer.name"
+const tag_net_sock_peer_addr_key string = "net.sock.peer.addr"
+
+const tag_messaging_system_value string = "rabbitmq"
+const tag_messaging_destination_kind_value string = "queue"
+const tag_messaging_protocol_value string = "AMQP"
+const tag_messaging_protocol_version_value string = "0.9.1"
+
+const tag_net_sock_family_value string = "inet"
